src/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==3.6.4.0
black==22.6.0
boto3==1.34.70
botocore==1.34.110
CacheControl==0.14.0
cachetools==5.3.3
celery==5.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coolsms-python-sdk==2.0.3
cryptography==42.0.7
Django==4.2.3
django-ckeditor==6.4.2
django-cors-headers==3.10.0
django-extensions==3.2.3
django-filter==21.1
django-hosts==5.1
django-js-asset==2.2.0
django-ordered-model==3.7.4
django-safedelete==1.3.1
django-storages==1.12.3
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.27.2
firebase-admin==6.5.0
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
gunicorn==20.1.0
httplib2==0.22.0
idna==3.7
inflection==0.5.1
isort==5.10.1
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
msgpack==1.0.8
mypy-extensions==1.0.0
pathspec==0.12.1
Pillow==9.2.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycurl==7.44.1
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.2
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.1
sentry-sdk==1.6.0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
watchtower==3.0.0
wcwidth==0.2.13
